requirements.txt
Django>=3.2.6,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.9.1,<2.10.0
Pillow>=8.3.1,<8.4.0
django-cleanup>=5.2.0,<5.3.0
djangorestframework-simplejwt>=4.7.2,<4.8.0
gunicorn>=20.1.0,<20.2.0
drf-yasg>=1.20.0,<1.21.0
flake8>=3.9.2,<3.10.0